+++ CHANGELOG.md
@@ -10,11 +10,16 @@ and this project adheres to [Semantic Versioning][semver].
 
 ### Added
 
+- Add key names from config files to metadata during repository setup, following updates to TUF and securesystemslib [(583)]
+- Implement iteration over all inserted YubiKeys during metadata signing [(583)]
+- Implement a `PinManager` class to allow secure pin reuse across API functions and eliminated insecure global pin storage [(583)]
 
 ### Changed
 
 
 ### Fixed
+
+[583]: https://github.com/openlawlibrary/taf/pull/583
 
 ## [0.34.1] - 02/11/2025
 
@@ -42,7 +47,7 @@ and this project adheres to [Semantic Versioning][semver].
 
 ### Changed
 
-Transition to the newest version of TUF [(561)]
+- Transition to the newest version of TUF [(561)]
 
 ### Fixed